Add booking by id

+++ services/nailserver/src/main/java/com/lovelyglam/nailserver/service/BookingService.java
@@ -6,7 +6,6 @@ import java.sql.Timestamp;
 import java.util.Collection;
 
 import com.lovelyglam.database.model.dto.request.SearchRequestParamsDto;
-import com.lovelyglam.database.model.dto.response.BookingPaymentResponse;
 import com.lovelyglam.database.model.dto.response.BookingResponse;
 import com.lovelyglam.database.model.dto.response.PaginationResponse;
 
@@ -14,6 +13,5 @@ public interface BookingService {
     PaginationResponse<BookingResponse> getBookingsByShopId(SearchRequestParamsDto request);
     Collection<BookingResponse> getBookingsByDayAndShopId(Date makingDate);
     Collection<BookingResponse> getBookingsByStartTimeAndShopId(Timestamp startTime);
-    BookingResponse getBookingById(Long id);
-
+    BookingResponse getBookingById(BigDecimal id);
 }
